Add support for default search field
The `q` query parameter searches the default search field (which, by
default, is all text fields), but you can acheive the same in a boolean
query (`bq` parameter) by NOT specifying a field for a text search. This
can be important because searching several text fields using the 'or'
operator does weird things to text relevance scores.