Add retry with exp. backoff on failed requests
Exponential backoff starts at 1 second and goes by powers of 2, so the first
retry is at 1s, the next at 3s, then 7s, then 15s, and so on. By default, it's
truncated at 3 retries and only attempted when the HTTP status code for a
failed request is 507 or 509. The number of retries and the applicable status
codes can be configured when creating a client.